+++ redwood-tooling-codegen/src/main/kotlin/app/cash/redwood/tooling/codegen/codegen.kt
@@ -57,7 +57,6 @@ internal fun SchemaSet.generateFileSpecs(type: CodegenType): List<FileSpec> {
       }
 
       Testing -> {
-        add(testingDeprecations(schema))
         add(generateTester(this@generateFileSpecs))
         add(generateMutableWidgetFactory(schema))
         for (widget in schema.widgets) {