+++ redwood-tooling-codegen/src/main/kotlin/app/cash/redwood/tooling/codegen/protocolCodegen.kt
@@ -36,7 +36,6 @@ internal fun ProtocolSchemaSet.generateFileSpecs(type: ProtocolCodegenType): Lis
   return buildList {
     when (type) {
       Guest -> {
-        add(protocolGuestDeprecations(schema))
         add(generateProtocolBridge(this@generateFileSpecs))
         add(generateComposeProtocolModifierSerialization(this@generateFileSpecs))
         for (dependency in all) {
@@ -49,7 +48,6 @@ internal fun ProtocolSchemaSet.generateFileSpecs(type: ProtocolCodegenType): Lis
       }
 
       Host -> {
-        add(protocolHostDeprecations(schema))
         add(generateProtocolFactory(this@generateFileSpecs))
         for (dependency in all) {
           generateProtocolModifierImpls(dependency, host = schema)?.let { add(it) }